requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
WTForms==3.1.2
WeasyPrint==66.0
pydyf==0.8.0
